feat(navbar): add cart icon to navbar
This commit updates the navbar component to include a cart icon that displays the number of items currently in the users cart.

The implementation of the cart icon involves updating the navbar component to include a computed property that retrieves the current number of items in the users cart, and updates the icons display accordingly. The icon also includes a `router-link` that takes the user to the CartView page when clicked.

Fixed a error too in quantity items when user try to add some product to the cart.

Closes #AdooreiTeste013